Испавление проблемы с делением

+++ task1/2.3/calc.c
@@ -13,10 +13,10 @@ int multiply(int a, int b) {
     return a * b;
 }
 
-int divide(int a, int b) {
+double divide(int a, int b) {
     if (b == 0) {
         printf("Ошибка: Нельзя делить на ноль.\n");
-        return 0;
+        return 0.0;
     }
-    return a / b;
+    return (double)a / (double)b;
 }
